requirements.txt
feedparser
requests
beautifulsoup4
lxml
python-dateutil

google-api-python-client
google-auth
google-auth-httplib2

pandas
